fix(stella-core): let the effective-budget latch adopt the first corrected calibration value
The latch (#1841) captured the compaction budget the first time the model
was known — which on a fresh session is one drift sample into a
three-sample warm-up, so it froze the 1.0 identity for the whole turn.
Every trial of bench match cc00894779ff reported calibration_factor: 1.0
on every step while estimates ran 2.4-9x under provider actuals, and every
compaction decision ran on a fraction of the real context: the 40+ samples
the turn itself recorded were never read back into any decision.

The identity factor is the estimator's still-warming answer (and the only
factor under which the budget equals the raw configured one), so it is now
served live and never captured; the first corrected value is what settles.
This keeps #1841's guarantee — the identity cannot move, so the budget
still changes at most once mid-turn, at the settling step, then freezes.

Witnesses (fail on the old latch, pass now — flip verified by reverting
the latch condition alone):
- step::tests::a_warming_calibrations_identity_is_served_live_never_captured
- driver::tests::calibration::a_fresh_sessions_calibration_factor_leaves_identity_within_the_turn

The #1841 witness now starts from a corrected value and is renamed
the_compaction_budget_is_latched_once_the_calibration_corrects. The
token-drift calibration section of driver/tests.rs moves unchanged to a
new driver/tests/calibration.rs submodule (the settlement.rs split
pattern): tests.rs is a grandfathered god file with zero headroom, and
the new end-to-end witness needed a home.

Closes #2133
